refatoracao do codigo da oh do Gabiz
