+++ src/services/firebase.js
@@ -31,113 +31,11 @@ const signUp = async (name, email, password) => {
 
 const forgotYourPassword = (email) => firebase.auth().sendPasswordResetEmail(email);
 
-
 // FEED
 
 const loadPosts = () => {
   const postsCollection = firebase.firestore().collection('posts');
-  postsCollection.orderBy('data','desc').get().then((snap) => {
-    document.querySelector('.posts-list').innerHTML = '';
-    snap.forEach((post) => {
-      addPosts(post);
-    });
-  });
-};
-
-const addPosts = (post) => {
-  const postTemplate = `
-    <li id="${post.data().userId}" data-template class="post-container">
-
-        <div class="user-info-container">
-          <!-- <img src="https://i.pravatar.cc/100?img=48" alt="User Photo" class="user-post-photo"> -->
-          <p class="user-name">@${post.data().userName}</p>
-          <p class="post-date" id="">${post.data().data}</p>
-        </div>
-
-        <div class="post-field">
-          <p class="user-post">${post.data().text}</p>
-
-          <div data-editcontainer class="edit-container display-none">
-            <textarea data-text="${post.id}" class="edit-post-textarea" rows="3" cols="30">${post.data().text}</textarea>
-
-            <div class="edit-buttons">
-              <button data-cancel="${post.id}" class="manage-edit-btn cancel-btn">Cancelar</button>
-              <button data-save="${post.id}" class="manage-edit-btn save-btn">Salvar</button>
-            </div>
-          </div>
-        </div>
-
-        <div class="manage-post" id=${post.id}>
-          <div class="post-likes" id="${post.id}">
-            <button id="like-btn" class="manage-post-btn like-btn"><i class="fas fa-heart" id="heart"></i></button>
-            <p class="likes-number" id="${post.id}">${post.data().likes}</p>
-          </div>
-          <button class="manage-post-btn edit-btn"><i data-edit="${post.id}" class="fas fa-pencil-alt"></i></button>
-          <button class="manage-post-btn delete-btn"><i class="fas fa-trash-alt"></i></button>
-        </div>
-     </li>
-   `;
-
-  document.querySelector('#postsList').innerHTML += postTemplate;
-  
-  const postsListContainer = document.querySelector('#postsList');
-
-  //função editar post
-  postsListContainer.addEventListener('click', (e) => {
-    const { target } = e;
-    const editPostButton = target.dataset.edit;
-    const cancelEditionButton = target.dataset.cancel; 
-    const saveEditionButton = target.dataset.save;
-
-    //Open edit
-    if(editPostButton) {
-      const editPostContainer = target.parentNode.parentNode.parentNode.querySelector('.edit-container')
-      const userPost = target.parentNode.parentNode.parentNode.querySelector('.user-post');
-
-      editPostContainer.classList.toggle('display-none');
-      userPost.classList.toggle('display-none');
-    }
-    //cancel edit
-    if(cancelEditionButton) {
-      const liElement = target.parentNode.parentNode.parentNode.parentNode;
-      const userPost = liElement.querySelector('.user-post');
-      const editcontainer = liElement.querySelector('.edit-container');
-
-      editcontainer.classList.toggle('display-none')
-      userPost.classList.toggle('display-none');
-    }
-    //save edit
-    if(saveEditionButton){
-      const liElement = target.parentNode.parentNode.parentNode.parentNode;
-      const textArea = liElement.querySelector('.edit-post-textarea');
-      const newText = textArea.value;
-      const postId = textArea.dataset.text;
-    
-      editPost(newText, postId)
-      document.querySelector('#postsList').innerHTML = '';
-      loadPosts()
-    }
-  });
- 
-  //função excluir posts
-  const deleteButtons = document.querySelectorAll('.delete-btn')
-  for (const button of deleteButtons) {
-    button.addEventListener('click', (e) => {
-      e.preventDefault()
-      deletePost(e.currentTarget.parentNode.id)
-      document.querySelector('#postsList').innerHTML = ''
-    });
-  };
-
-  //função like posts
-  const likeButtons = document.querySelectorAll('.like-btn');
-  for (const button of likeButtons) {
-    button.addEventListener('click', (e) => {
-      e.preventDefault()
-      likePosts(e.currentTarget.parentNode.id)   
-      document.querySelector('#postsList').innerHTML = ''
-    });
-  }
+  return postsCollection.orderBy('data', 'desc').get();
 };
 
 const editPost = (newText, postID) => {
@@ -153,38 +51,35 @@ const deletePost = (postId) => {
 };
 
 const likePosts = (postId) => {
-  const postsCollection = firebase.firestore().collection("posts");
+  const postsCollection = firebase.firestore().collection('posts');
   const promisseResult = postsCollection.doc(postId).get()
-      .then((post => {
-        console.log(postId)
-        console.log(post.data())
-        const countLikes = post.data().likes;
-        if(countLikes >= 1) {
-            postsCollection.doc(postId).update({
-              likes: post.data().likes - 1
-            })
-            .then(() => {
-              loadPosts();
-            });
-        } else {
-          postsCollection.doc(postId).update({
-            likes: post.data().likes + 1
-          })
+    .then(((post) => {
+      console.log(postId);
+      console.log(post.data());
+      const countLikes = post.data().likes;
+      if (countLikes >= 1) {
+        postsCollection.doc(postId).update({
+          likes: post.data().likes - 1,
+        })
           .then(() => {
             loadPosts();
           });
-        }
-      }))
-      return promisseResult.then();
+      } else {
+        postsCollection.doc(postId).update({
+          likes: post.data().likes + 1,
+        })
+          .then(() => {
+            loadPosts();
+          });
+      }
+    }));
+  return promisseResult.then();
 };
-
-
 
 const postData = () => {
   const data = new Date();
   return data.toLocaleString('pt-BR');
 };
-
 
 const createPost = (textPost) => {
   const user = firebase.auth().currentUser;
@@ -201,13 +96,8 @@ const createPost = (textPost) => {
   const postsCollection = firebase
     .firestore()
     .collection('posts');
-  postsCollection.add(post).then(() => {
-    document.querySelector('#postsList').innerHTML = '';
-    loadPosts();
-  });
+  return postsCollection.add(post);
 };
-
-
 
 export default {
   waitAuthState,
@@ -220,4 +110,8 @@ export default {
   forgotYourPassword,
   createPost,
   loadPosts,
+  editPost,
+  deletePost,
+  likePosts,
+
 };
